Update tests using multidimensional meshes.
Add a test for ghost layers in MarchingCubes.
Combine ghost option from gen-strided-structured-mesh.py
into gen-multidom-structured-mesh.py.  Regenerated
data files.
