+++ src/axom/quest/examples/CMakeLists.txt
@@ -231,10 +231,12 @@ if(CONDUIT_FOUND)
         endif()
 
         # Non-zero empty-rank probability tests domain underloading case
-        set(_meshes "mdmesh.2x1" "mdmesh.2x3" "mdmesh.2x2x1")
+        set(_meshes "mdmesh.2x1" "mdmesh.2x3" "mdmesh.2x2x1" "mdmeshg.2x2x1")
         # The amc.* files were generated by these commands:
-        # src/tools/gen-multidom-structured-mesh.py -ml=0,0 -mu=2,2 -ms=20,20 -dc=2,2 -o amc_mesh.2x2
+        # src/tools/gen-multidom-structured-mesh.py -ml=0,0 -mu=2,2 -ms=100,100 -dc=2,1 -o mdmesh.2x1
+        # src/tools/gen-multidom-structured-mesh.py -ml=0,0 -mu=2,2 -ms=100,100 -dc=2,3 -o mdmesh.2x3
         # src/tools/gen-multidom-structured-mesh.py -ml=0,0,0 -mu=2,2,2 -ms=20,20,15 -dc=2,2,1 -o mdmesh.2x2x1
+        # src/tools/gen-multidom-structured-mesh.py -ml=0,0,0 -mu=2,2,2 -ms=20,20,15 -dc=2,2,1 -o mdmeshg.2x2x1 --strided
         foreach(_pol ${_policies})
             foreach(_mesh ${_meshes})
                 # Determine problem dimension by mesh filename.
